# Use a clock-based Buffer Manager eviction strategy