--- a/src/storage/buffer_manager/buffer_manager.cpp
+++ b/src/storage/buffer_manager/buffer_manager.cpp
@@ -1,9 +1,12 @@
 #include "storage/buffer_manager/buffer_manager.h"
 
+#include <atomic>
 #include <cstring>
 
+#include "common/assert.h"
 #include "common/constants.h"
 #include "common/exception/buffer_manager.h"
+#include "storage/buffer_manager/bm_file_handle.h"
 
 #if defined(_WIN32)
 #include <exception>
@@ -17,61 +20,58 @@ using namespace kuzu::common;
 
 namespace kuzu {
 namespace storage {
-// In this function, we try to remove as many as possible candidates that are not evictable from the
-// eviction queue until we hit a candidate that is evictable.
-// 1) If the candidate page's version has changed, which means the page was pinned and unpinned, we
-// remove the candidate from the queue.
-// 2) If the candidate page's state is UNLOCKED, and its page version hasn't changed, which means
-// the page was optimistically read, we give a second chance to evict the page by marking the page
-// as MARKED, and moving the candidate to the back of the queue.
-// 3) If the candidate page's state is LOCKED, we remove the candidate from the queue.
-void EvictionQueue::removeNonEvictableCandidates() {
-    std::shared_lock sLck{mtx};
-    while (true) {
-        EvictionCandidate evictionCandidate;
-        if (!queue->try_dequeue(evictionCandidate)) {
-            break;
+
+bool EvictionQueue::insert(uint32_t fileIndex, common::page_idx_t pageIndex, PageState* pageState) {
+    EvictionCandidate candidate{fileIndex, pageIndex, pageState};
+    while (size < capacity) {
+        // Weak is fine since spurious failure is acceptable.
+        // The slot can always be filled later.
+        auto emptyCandidate = EMPTY;
+        if (data[insertCursor.fetch_add(1, std::memory_order_relaxed) % capacity]
+                .compare_exchange_weak(emptyCandidate, candidate)) {
+            size++;
+            return true;
         }
-        auto pageStateAndVersion = evictionCandidate.pageState->getStateAndVersion();
-        if (evictionCandidate.isEvictable(pageStateAndVersion)) {
-            queue->enqueue(evictionCandidate);
-            break;
-        } else if (evictionCandidate.isSecondChanceEvictable(pageStateAndVersion)) {
-            // The page was optimistically read, mark it as MARKED, and enqueue to be evicted later.
-            evictionCandidate.pageState->tryMark(pageStateAndVersion);
-            queue->enqueue(evictionCandidate);
-            continue;
-        } else {
-            // Cases to remove the candidate from the queue:
-            // 1) The page is currently LOCKED (it is currently pinned), remove the candidate from
-            // the queue.
-            // 2) The page's version number has changed (it was pinned and unpinned), another
-            // candidate exists for this page in the queue. remove the candidate from the queue.
-            continue;
+    }
+    return false;
+}
+
+std::atomic<EvictionCandidate>* EvictionQueue::next() {
+    std::atomic<EvictionCandidate>* candidate;
+    do {
+        // Since the buffer pool size is a power of two (as is the page size), (UINT64_MAX + 1) %
+        // size == 0, which means no entries will be skipped when the cursor overflows
+        candidate = &data[evictionCursor.fetch_add(1, std::memory_order_relaxed) % capacity];
+    } while (candidate->load() == EMPTY);
+    return candidate;
+}
+
+void EvictionQueue::removeCandidatesForFile(uint32_t fileIndex) {
+    for (uint64_t i = 0; i < capacity; i++) {
+        auto candidate = data[i].load();
+        if (candidate.fileIdx == fileIndex && data[i].compare_exchange_strong(candidate, EMPTY)) {
+            size--;
         }
     }
 }
 
-void EvictionQueue::removeCandidatesForFile(kuzu::storage::BMFileHandle& fileHandle) {
-    std::unique_lock xLck{mtx};
-    EvictionCandidate candidate;
-    uint64_t loopedCandidateIdx = 0;
-    auto numCandidatesInQueue = queue->size();
-    while (loopedCandidateIdx < numCandidatesInQueue && queue->try_dequeue(candidate)) {
-        if (candidate.fileHandle != &fileHandle) {
-            queue->enqueue(candidate);
-        }
-        loopedCandidateIdx++;
+void EvictionQueue::clear(std::atomic<EvictionCandidate>& candidate) {
+    auto nonEmpty = candidate.load();
+    if (nonEmpty != EMPTY && candidate.compare_exchange_strong(nonEmpty, EMPTY)) {
+        size--;
+        return;
     }
+    KU_UNREACHABLE;
 }
 
 BufferManager::BufferManager(uint64_t bufferPoolSize, uint64_t maxDBSize)
-    : usedMemory{0}, bufferPoolSize{bufferPoolSize}, numEvictionQueueInsertions{0} {
+    : bufferPoolSize{bufferPoolSize},
+      evictionQueue{bufferPoolSize / BufferPoolConstants::PAGE_4KB_SIZE},
+      usedMemory{evictionQueue.getCapacity() * sizeof(EvictionCandidate)} {
     verifySizeParams(bufferPoolSize, maxDBSize);
     vmRegions.resize(2);
     vmRegions[0] = std::make_unique<VMRegion>(PageSizeClass::PAGE_4KB, maxDBSize);
     vmRegions[1] = std::make_unique<VMRegion>(PageSizeClass::PAGE_256KB, bufferPoolSize);
-    evictionQueue = std::make_unique<EvictionQueue>();
 }
 
 void BufferManager::verifySizeParams(uint64_t bufferPoolSize, uint64_t maxDBSize) {
@@ -109,6 +109,10 @@ uint8_t* BufferManager::pin(BMFileHandle& fileHandle, page_idx_t pageIdx,
                 if (!claimAFrame(fileHandle, pageIdx, pageReadPolicy)) {
                     pageState->unlock();
                     throw BufferManagerException("Failed to claim a frame.");
+                }
+                if (!evictionQueue.insert(fileHandle.getFileIndex(), pageIdx, pageState)) {
+                    throw BufferManagerException(
+                        "Eviction queue is full! This should be impossible.");
                 }
                 return getFrame(fileHandle, pageIdx);
             }
@@ -228,7 +232,39 @@ void BufferManager::optimisticRead(BMFileHandle& fileHandle, page_idx_t pageIdx,
 void BufferManager::unpin(BMFileHandle& fileHandle, page_idx_t pageIdx) {
     auto pageState = fileHandle.getPageState(pageIdx);
     pageState->unlock();
-    addToEvictionQueue(&fileHandle, pageIdx, pageState);
+}
+
+// evicts up to 64 pages and returns the space reclaimed
+uint64_t BufferManager::evictPages() {
+    const size_t BATCH_SIZE = 64;
+    std::array<std::atomic<EvictionCandidate>*, BATCH_SIZE> evictionCandidates;
+    size_t pagesEvicted = 0;
+    size_t pagesTried = 0;
+    uint64_t claimedMemory = 0;
+
+    // Try each page at least twice.
+    // E.g. if the vast majority of pages are unmarked and unlocked,
+    // the first pass will mark them and the second pass, if insufficient marked pages
+    // are found, will evict the first batch.
+    auto failureLimit = evictionQueue.getSize() * 2;
+    while (pagesEvicted < BATCH_SIZE && pagesTried < failureLimit) {
+        evictionCandidates[pagesEvicted] = evictionQueue.next();
+        pagesTried++;
+        auto evictionCandidate = evictionCandidates[pagesEvicted]->load();
+        auto pageStateAndVersion = evictionCandidate.pageState->getStateAndVersion();
+        if (!evictionCandidate.isEvictable(pageStateAndVersion)) {
+            if (evictionCandidate.isSecondChanceEvictable(pageStateAndVersion)) {
+                evictionCandidate.pageState->tryMark(pageStateAndVersion);
+            }
+            continue;
+        }
+        pagesEvicted++;
+    }
+
+    for (size_t i = 0; i < pagesEvicted; i++) {
+        claimedMemory += tryEvictPage(*evictionCandidates[i]);
+    }
+    return claimedMemory;
 }
 
 // This function tries to load the given page into a frame. Due to our design of mmap, each page is
@@ -252,52 +288,34 @@ bool BufferManager::claimAFrame(BMFileHandle& fileHandle, page_idx_t pageIdx,
 bool BufferManager::reserve(uint64_t sizeToReserve) {
     // Reserve the memory for the page.
     auto currentUsedMem = reserveUsedMemory(sizeToReserve);
-    uint64_t claimedMemory = 0;
+    uint64_t totalClaimedMemory = 0;
     // Evict pages if necessary until we have enough memory.
-    while ((currentUsedMem + sizeToReserve - claimedMemory) > bufferPoolSize.load()) {
-        EvictionCandidate evictionCandidate;
-        if (!evictionQueue->dequeue(evictionCandidate)) {
+    while ((currentUsedMem + sizeToReserve - totalClaimedMemory) > bufferPoolSize.load()) {
+        auto memoryClaimed = evictPages();
+        if (memoryClaimed == 0) {
             // Cannot find more pages to be evicted. Free the memory we reserved and return false.
-            freeUsedMemory(sizeToReserve + claimedMemory);
+            freeUsedMemory(sizeToReserve + totalClaimedMemory);
             return false;
         }
-        auto pageStateAndVersion = evictionCandidate.pageState->getStateAndVersion();
-        if (!evictionCandidate.isEvictable(pageStateAndVersion)) {
-            if (evictionCandidate.isSecondChanceEvictable(pageStateAndVersion)) {
-                evictionCandidate.pageState->tryMark(pageStateAndVersion);
-                evictionQueue->enqueue(evictionCandidate);
-            }
-            continue;
-        }
-        // We found a page that potentially hasn't been accessed since enqueued. We try to evict the
-        // page from its frame by calling `tryEvictPage`, which will check if the page's version has
-        // changed, if not, we evict the page from its frame.
-        claimedMemory += tryEvictPage(evictionCandidate);
+        totalClaimedMemory += memoryClaimed;
         currentUsedMem = usedMemory.load();
     }
-    if ((currentUsedMem + sizeToReserve - claimedMemory) > bufferPoolSize.load()) {
+    if ((currentUsedMem + sizeToReserve - totalClaimedMemory) > bufferPoolSize.load()) {
         // Cannot claim the memory needed. Free the memory we reserved and return false.
-        freeUsedMemory(sizeToReserve + claimedMemory);
+        freeUsedMemory(sizeToReserve + totalClaimedMemory);
         return false;
     }
     // Have enough memory available now
-    freeUsedMemory(claimedMemory);
+    freeUsedMemory(totalClaimedMemory);
     return true;
 }
 
-void BufferManager::addToEvictionQueue(BMFileHandle* fileHandle, page_idx_t pageIdx,
-    PageState* pageState) {
-    auto currStateAndVersion = pageState->getStateAndVersion();
-    if (++numEvictionQueueInsertions == BufferPoolConstants::EVICTION_QUEUE_PURGING_INTERVAL) {
-        evictionQueue->removeNonEvictableCandidates();
-        numEvictionQueueInsertions = 0;
+uint64_t BufferManager::tryEvictPage(std::atomic<EvictionCandidate>& _candidate) {
+    auto candidate = _candidate.load();
+    // Page must have been evicted by another thread already
+    if (candidate.pageState == nullptr) {
+        return 0;
     }
-    pageState->tryMark(currStateAndVersion);
-    evictionQueue->enqueue(fileHandle, pageIdx, pageState,
-        PageState::getVersion(currStateAndVersion));
-}
-
-uint64_t BufferManager::tryEvictPage(EvictionCandidate& candidate) {
     auto& pageState = *candidate.pageState;
     auto currStateAndVersion = pageState.getStateAndVersion();
     // We check if the page is evictable again. Note that if the page's state or version has
@@ -305,12 +323,21 @@ uint64_t BufferManager::tryEvictPage(EvictionCandidate& candidate) {
     if (!candidate.isEvictable(currStateAndVersion) || !pageState.tryLock(currStateAndVersion)) {
         return 0;
     }
-    // At this point, the page is LOCKED. Next, flush out the frame into the file page if the frame
+    // The pageState was locked, but another thread already evicted this candidate and unlocked it
+    // before the lock occurred
+    if (_candidate.load() != candidate) {
+        pageState.unlock();
+        return 0;
+    }
+    // At this point, the page is LOCKED, and we have exclusive access to the eviction candidate.
+    // Next, flush out the frame into the file page if the frame
     // is dirty. Finally remove the page from the frame and reset the page to EVICTED.
-    flushIfDirtyWithoutLock(*candidate.fileHandle, candidate.pageIdx);
-    auto numBytesFreed = candidate.fileHandle->getPageSize();
-    releaseFrameForPage(*candidate.fileHandle, candidate.pageIdx);
+    auto* fileHandle = fileHandles[candidate.fileIdx];
+    flushIfDirtyWithoutLock(*fileHandle, candidate.pageIdx);
+    auto numBytesFreed = fileHandle->getPageSize();
+    releaseFrameForPage(*fileHandle, candidate.pageIdx);
     pageState.resetToEvicted();
+    evictionQueue.clear(_candidate);
     return numBytesFreed;
 }
 
@@ -334,7 +361,7 @@ void BufferManager::flushIfDirtyWithoutLock(BMFileHandle& fileHandle, page_idx_t
 }
 
 void BufferManager::removeFilePagesFromFrames(BMFileHandle& fileHandle) {
-    evictionQueue->removeCandidatesForFile(fileHandle);
+    evictionQueue.removeCandidatesForFile(fileHandle.getFileIndex());
     for (auto pageIdx = 0u; pageIdx < fileHandle.getNumPages(); ++pageIdx) {
         removePageFromFrame(fileHandle, pageIdx, false /* do not flush */);
     }