--- a/tools/rust_api/build.rs
+++ b/tools/rust_api/build.rs
@@ -32,6 +32,7 @@ fn link_libraries() {
             println!("cargo:rustc-link-lib=dylib=c++");
         } else {
             println!("cargo:rustc-link-lib=dylib=stdc++");
+            println!("cargo:rustc-link-lib=dylib=atomic");
         }
 
         println!("cargo:rustc-link-lib=static=utf8proc");