@ray6080 2024-06-10
Can u also benchmark on some pure scan queries on large dataset like ldbc100 comment table? I think the cold run should be able to present performance difference.