@benjaminwinger 2024-06-11
> I feel we should start to take the mem usage of the queue into consideration. BM's available memory should be bufferPoolSize - queueSize. Though the ratio of queue's memory usage is small, the principle is to try to make memory usage more accurate when possible and convenient. What do you think?

I think that's reasonable.