@benjaminwinger 2024-06-11
> Can u also benchmark on some pure scan queries on large dataset like ldbc100 comment table? I think the cold run should be able to present performance difference.

There is a difference, but it doesn't appear to be huge.

`MATCH (c:Comment) RETURN c.*;`:
- Before: 11.3 seconds 
- After: 10.7 seconds

`MATCH (c:Comment) RETURN max(c.content);`:
- Before: 4.9 seconds (6.0 seconds with 64MB buffer pool size)
- After: 4.5 seconds (5.5 seconds with 64MB buffer pool size)

(both run cold from a database stored on an ssd).